+++ packages/contracts/contracts/core/DAO.sol
@@ -14,7 +14,7 @@ import "./component/CallbackHandler.sol";
 import "./permission/PermissionManager.sol";
 import "./IDAO.sol";
 
-/// @title The public interface of the Aragon DAO framework.
+/// @title DAO
 /// @author Aragon Association - 2021
 /// @notice This contract is the entry point to the Aragon DAO framework and provides our users a simple and easy to use public interface.
 /// @dev Public API of the Aragon DAO framework.