+++ packages/contracts/contracts/core/IDAO.sol
@@ -6,7 +6,7 @@ pragma solidity 0.8.10;
 /// @author Aragon Association - 2022
 /// @notice The interface required for DAOs within the Aragon App DAO framework.
 abstract contract IDAO {
-    /// @notice The `IDAO` interface ID.
+    /// @notice The [ERC-165](https://eips.ethereum.org/EIPS/eip-165) interface ID of the contract.
     bytes4 internal constant DAO_INTERFACE_ID = type(IDAO).interfaceId;
 
     struct Action {
@@ -50,7 +50,7 @@ abstract contract IDAO {
     /// @param actor The address of the caller.
     /// @param callId The id of the call.
     /// @dev The value of callId is defined by the component/contract calling the execute function.
-    ///      A Component implementation can use it, for example, as a nonce.
+    ///      A `Plugin` implementation can use it, for example, as a nonce.
     /// @param actions Array of actions executed.
     /// @param execResults Array with the results of the executed actions.
     event Executed(address indexed actor, uint256 callId, Action[] actions, bytes[] execResults);