+++ packages/contracts/contracts/core/component/DaoAuthorizableCloneable.sol
@@ -12,7 +12,7 @@ import {DaoAuthorizableBase} from "./DaoAuthorizableBase.sol";
 /// @notice An abstract contract providing a meta-transaction compatible modifier for clonable contracts to authorize function calls through an associated DAO.
 /// @dev Make sure to call `__DaoAuthorizableCloneable_init` during initialization of the inheriting contract.
 ///      This contract is compatible with meta transactions through OZ's `Context`.
-abstract contract DaoAuthorizableCloneable is DaoAuthorizableBase, Initializable {
+abstract contract DaoAuthorizableCloneable is Initializable, DaoAuthorizableBase {
     /// @notice Initializes the contract by setting the associated DAO.
     /// @param _dao The associated DAO address.
     function __DaoAuthorizableCloneable_init(IDAO _dao) internal onlyInitializing {