+++ packages/contracts/contracts/plugin/PluginRepo.sol
@@ -1,23 +1,22 @@
-/*
- * SPDX-License-Identifier:    MIT
- */
+// SPDX-License-Identifier:    MIT
 
 pragma solidity 0.8.10;
 
 import {Initializable} from "@openzeppelin/contracts-upgradeable/proxy/utils/Initializable.sol";
+import {ERC165Upgradeable} from "@openzeppelin/contracts-upgradeable/utils/introspection/ERC165Upgradeable.sol";
+import {ERC165} from "@openzeppelin/contracts/utils/introspection/ERC165.sol";
 import {UUPSUpgradeable} from "@openzeppelin/contracts/proxy/utils/UUPSUpgradeable.sol";
 import {Address} from "@openzeppelin/contracts/utils/Address.sol";
-import {ERC165} from "@openzeppelin/contracts/utils/introspection/ERC165.sol";
-import {PermissionManager} from "../core/permission/PermissionManager.sol";
-import {ERC165Upgradeable} from "@openzeppelin/contracts-upgradeable/utils/introspection/ERC165Upgradeable.sol";
 
+import {PermissionManager} from "../core/permission/PermissionManager.sol";
 import {_uncheckedIncrement} from "../utils/UncheckedMath.sol";
 import {PluginSetup} from "./PluginSetup.sol";
 import {IPluginRepo} from "./IPluginRepo.sol";
 
 /// @title PluginRepo
 /// @author Aragon Association - 2020 - 2022
-/// @notice The repository contract required for managing and publishing different plugin versions within the Aragon DAO framework following the [Semantic Versioning 2.0.0](https://semver.org/) convention.
+/// @notice The plugin repository contract required for managing and publishing different plugin versions within the Aragon DAO framework following the [Semantic Versioning 2.0.0](https://semver.org/) convention.
+//TODO Rename to PluginSetupRepo?
 contract PluginRepo is
     IPluginRepo,
     Initializable,
@@ -213,7 +212,7 @@ contract PluginRepo is
     /// @notice Checks if a version bump is valid.
     /// @param _oldVersion The old semantic version number.
     /// @param _newVersion The new semantic version number.
-    /// @return bool True if the bump is valid.
+    /// @return bool Returns true if the bump is valid.
     function isValidBump(uint16[3] memory _oldVersion, uint16[3] memory _newVersion)
         public
         pure