+++ packages/contracts/contracts/plugin/examples/count/CounterV1.sol
@@ -6,29 +6,41 @@ import {PluginUUPSUpgradeable} from "../../../core/plugin/PluginUUPSUpgradeable.
 import {MultiplyHelper} from "./MultiplyHelper.sol";
 import {IDAO} from "../../../core/IDAO.sol";
 
-/// @notice The first version of example plugin - CounterV1.
+/// @title CounterV1
+/// @author Aragon Association - 2022
+/// @notice The first version of an example plugin counting numbers.
 contract CounterV1 is PluginUUPSUpgradeable {
+    /// @notice The ID of the permission required to call the `multiply` function.
     bytes32 public constant MULTIPLY_PERMISSION_ID = keccak256("MULTIPLY_PERMISSION");
 
+    /// @notice A counter varaible.
     uint256 public count;
+
+    /// @notice A helper contract associated with the plugin.
     MultiplyHelper public multiplyHelper;
 
+    /// @notice Initializes the plugin.
+    /// @param _dao The contract of the associated DAO.
+    /// @param _multiplyHelper The helper contract associated with the plugin to multiply numbers.
+    /// @param _count The inital value of the counter.
     function initialize(
         IDAO _dao,
         MultiplyHelper _multiplyHelper,
-        uint256 _num
+        uint256 _count
     ) external initializer {
         __DaoAuthorizableUpgradeable_init(_dao);
 
-        count = _num;
+        count = _count;
         multiplyHelper = _multiplyHelper;
     }
 
-    function multiply(uint256 _a) public auth(MULTIPLY_PERMISSION_ID) returns (uint256) {
-        count = multiplyHelper.multiply(count, _a);
-        return count;
+    /// @notice Multiplies the count with a number.
+    /// @param _a The number to multiply the coun with.
+    function multiply(uint256 _a) public view auth(MULTIPLY_PERMISSION_ID) returns (uint256) {
+        return multiplyHelper.multiply(count, _a);
     }
 
+    /// @notice Executes something on the DAO.
     function execute() public {
         // IDAO dao = getDao();
         // In order to do this, Count needs permission on the dao (EXEC_ROLE)