+++ packages/contracts/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added `DaoAuthorizableBase` and `DaoAuthorizableConstructable` classes.
+- Added `DaoAuthorizableBase` class.
 - Added `DaoAuthorizableClonable` using OpenZepplin initialization.
 - Added mocks and tests for the `Plugin` and `PluginSetup` classes.
 - Added `PluginSetupProcessor` to be the main class processing `PluginSetup` contracts and applying permissions in the installing DAO.
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Refactored NatSpec comments and names for the contracts related to the `Plugin` and `PluginSetup`.
 - Renamed `PluginTransparentUpgradeable` to `PluginUpgradeable`.
 - Refactored `AdaptiveERC165` into an independent `CallbackHandler` contract and separated `ERC165` from it.
 - Adapted `Component` to use `DaoAuthorizableUpgradeable` until it is fully refactored to become `Plugin`.